discovery_service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-consul==1.1.0
httpx==0.25.2
structlog==23.2.0 